实现用户进程hello, exit

+++ kernel/driver/ide.c
@@ -4,12 +4,12 @@
 #include "defs.h"
 
 
-static byte ide[IDE_SIZE];
+static u8 ide[IDE_SIZE];
 
-void read_ide_sects(void* dst, uint32_t src_sect, size_t num_sects) {
+void read_ide_sects(void* dst, u32 src_sect, u64 num_sects) {
     memcpy(dst, &ide[src_sect * SECTION_SIZE], num_sects * SECTION_SIZE);
 }
 
-void write_ide_sects(const void* src, uint32_t dst_sect, size_t num_sects) {
+void write_ide_sects(const void* src, u32 dst_sect, u64 num_sects) {
     memcpy(&ide[dst_sect * SECTION_SIZE], src, num_sects * SECTION_SIZE);
 }
